+++ dfetch/commands/common.py
@@ -6,43 +6,43 @@ import yaml
 
 from dfetch.log import get_logger
 from dfetch.manifest.manifest import Manifest
-from dfetch.manifest.parse import get_childmanifests
+from dfetch.manifest.parse import get_submanifests
 from dfetch.manifest.project import ProjectEntry
 
 logger = get_logger(__name__)
 
 
-def check_child_manifests(manifest: Manifest, project: ProjectEntry) -> None:
-    """Check for child manifests within a project.
+def check_sub_manifests(manifest: Manifest, project: ProjectEntry) -> None:
+    """Check for sub-manifests within a project.
 
     Args:
         manifest (dfetch.manifest.manifest.Manifest): The parent manifest with projects.
         project (ProjectEntry): The parent project.
     """
-    for childmanifest in get_childmanifests(skip=[manifest.path]):
+    for submanifest in get_submanifests(skip=[manifest.path]):
         recommendations: list[ProjectEntry] = []
-        for childproject in childmanifest.projects:
-            if childproject.remote_url not in [
+        for subproject in submanifest.projects:
+            if subproject.remote_url not in [
                 project.remote_url for project in manifest.projects
             ]:
-                recommendations.append(childproject.as_recommendation())
+                recommendations.append(subproject.as_recommendation())
 
         if recommendations:
-            childmanifest_relpath = os.path.relpath(
-                childmanifest.path, start=os.path.dirname(manifest.path)
+            submanifest_relpath = os.path.relpath(
+                submanifest.path, start=os.path.dirname(manifest.path)
             ).replace("\\", "/")
-            _make_recommendation(project, recommendations, childmanifest_relpath)
+            _make_recommendation(project, recommendations, submanifest_relpath)
 
 
 def _make_recommendation(
-    project: ProjectEntry, recommendations: list[ProjectEntry], childmanifest_path: str
+    project: ProjectEntry, recommendations: list[ProjectEntry], submanifest_path: str
 ) -> None:
     """Make recommendations to the user.
 
     Args:
         project (ProjectEntry): The parent project.
         recommendations (List[ProjectEntry]): List of recommendations
-        childmanifest_path (str): Path to the source of recommendations
+        submanifest_path (str): Path to the source of recommendations
     """
     recommendation_json = yaml.dump(
         [proj.as_yaml() for proj in recommendations],
@@ -54,7 +54,7 @@ def _make_recommendation(
         "\n".join(
             [
                 f'"{project.name}" depends on the following project(s) which are not part of your manifest:',
-                f"(found in {childmanifest_path})",
+                f"(found in {submanifest_path})",
                 "",
                 recommendation_json,
                 "",